rename: 공통 타입 파일 위치 수정-#39
