+++ src/types/components/index.ts
@@ -1,0 +1,6 @@
+export type * from "./button";
+export type * from "./input";
+export type * from "./modal";
+export type * from "./select";
+export type * from "./textarea";
+export type * from "./title";
