--- src/types/contexts/calendarContext.ts
@@ -1,5 +1,0 @@
-export interface CalendarContextProps {
-	currentDate: Date;
-	onPrevMonth: () => void;
-	onNextMonth: () => void;
-}
