+++ src/types/index.ts
@@ -1,3 +1,2 @@
-export * from "./components/button";
-export type { default as PageTitleProps } from "./components/title";
-export type * from "./contexts/calendarContext";
+export type * from "./components";
+export type * from "./store";
