+++ src/types/store/user.ts
@@ -1,0 +1,7 @@
+export interface UserInfoState {
+	email: string | null;
+	name: string | null;
+	grade: string | null;
+	team: string | null;
+	photoURL: string | null;
+}
